Fit a gaussian profile, line emission IFU
Imagen_in= Data_cube_fitsfile. xyz (optimized for GMOS-IFU)
l_min_izq=  starting lambda of the continuum on the left of the emission line
l_max_izq= ending lambda of the continuum on the left of the emission line
l_min_der= starting lambda of the continuum on the right of the emission line
l_max_der= ending lambda of the continuum on the right of the emission line

guess_line= Guess of the central FWHM of the line
guess_FWHM= guess of the FWHM oof the line

Orden_pol= Polynomial applied to the background 

FILE_OUT= Text file with the results of the fits.

The code does 3 iteration per fitting. It usually converges on the second step